Port remaining tests to new TestBase base class.
