+++ tests/python/pants_test/binaries/BUILD
@@ -12,7 +12,6 @@ python_tests(
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test:test_base',
-    'tests/python/pants_test:base_test',
   ]
 )
